refactor(frontend): remove home pathname
