+++ frontend/__tests__/components/pages/Error404Page.test.tsx
@@ -4,7 +4,7 @@
 import '@testing-library/jest-dom'
 import { render, screen } from '@testing-library/react'
 
-import Error404Page from '../../../src/components/error-pages/Error404Page'
+import Error404Page from '../../../src/components/pages/Error404Page'
 
 describe('Error 404 Page', () => {
   it('renders 404 without crashing', () => {
