+++ frontend/__tests__/components/pages/ErrorPage.test.tsx
@@ -4,7 +4,7 @@
 import '@testing-library/jest-dom'
 import { render, screen } from '@testing-library/react'
 
-import ErrorPage from '../../../src/components/error-pages/ErrorPage'
+import ErrorPage from '../../../src/components/pages/ErrorPage'
 
 describe('Error Page', () => {
   it('renders custom statusCode without crashing', () => {
