+++ frontend/src/utils/next-i18next-utils.ts
@@ -28,7 +28,6 @@ export const getNamespaces = (
     return [...new Set([...namespaces, namespacesRequired])]
   }
 
-  namespacesRequired
   return [...new Set([...namespaces, ...namespacesRequired])]
 }
